ci: share optional diff artifact runner
